+++ src/components/Input/WriteInput/WriteInput.stories.tsx
@@ -1,0 +1,17 @@
+import type { Meta } from '@storybook/react';
+
+import WriteInput from '.';
+
+const meta: Meta<typeof WriteInput> = {
+  title: 'WriteInput',
+  component: WriteInput,
+  decorators: [
+    (Story) => (
+      <div style={{ height: '500px' }}>
+        <Story />
+      </div>
+    ),
+  ],
+};
+
+export default meta;
